requirements.txt
anthropic==0.34.0
feedparser==6.0.11
httpx==0.27.0
python-dotenv==1.0.1
python-telegram-bot==21.6
elevenlabs==1.9.0
pydub==0.25.1
imageio-ffmpeg==0.5.1
Pillow==10.4.0
openai==1.51.0
